+++ src/galdynamix/dynamics/__init__.py
@@ -1,0 +1,10 @@
+"""galdynamix: Galactic Dynamix in Jax"""
+# ruff: noqa: F403
+
+from __future__ import annotations
+
+from . import mockstream
+from .mockstream import *
+
+__all__: list[str] = []
+__all__ += mockstream.__all__
