+++ src/galdynamix/dynamics/mockstream/__init__.py
@@ -1,0 +1,12 @@
+"""galdynamix: Galactic Dynamix in Jax"""
+# ruff: noqa: F403
+
+from __future__ import annotations
+
+from . import _df, _mockstream
+from ._df import *
+from ._mockstream import *
+
+__all__: list[str] = []
+__all__ += _df.__all__
+__all__ += _mockstream.__all__
