+++ src/galdynamix/integrate/_builtin/leapfrog.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-__all__ = ["leapfrog_run"]
+__all__ = []  # WIP
 
 from functools import partial
 from typing import Callable, Protocol
