+++ src/galdynamix/potential/_potential/composite.py
@@ -3,8 +3,10 @@ from __future__ import annotations
 __all__ = ["CompositePotential"]
 
 
-from dataclasses import InitVar
+from collections.abc import Mapping
+from typing import TypeVar
 
+import equinox as eqx
 import jax.numpy as xp
 import jax.typing as jt
 
@@ -12,15 +14,59 @@ from galdynamix.utils import jit_method
 
 from .base import PotentialBase
 
+V = TypeVar("V")
+
+
+class FrozenDict(Mapping[str, V]):
+    def __init__(self, *args: V, **kwargs: V) -> None:
+        self._data: dict[str, V] = dict(*args, **kwargs)
+
+    def __getitem__(self, key: str) -> V:
+        return self._data[key]
+
+    def __iter__(self) -> iter[str]:
+        return iter(self._data)
+
+    def __len__(self) -> int:
+        return len(self._data)
+
+    def __hash__(self) -> int:
+        return hash(tuple(self._data.items()))
+
+    def keys(self) -> iter[str]:
+        return self._data.keys()
+
+    def values(self) -> iter[V]:
+        return self._data.values()
+
+    def items(self) -> iter[tuple[str, V]]:
+        return self._data.items()
+
 
 class CompositePotential(PotentialBase):
     """Composite Potential."""
 
-    potential_list: InitVar[list[PotentialBase]]
+    # potentials: FrozenDict[str, PotentialBase] = eqx.field(converter=FrozenDict)
 
-    def __post_init__(self, potential_list: list[PotentialBase]) -> None:
-        super().__post_init__()
-        object.__setattr__(self, "_potential_list", potential_list)
+    potentials: tuple[PotentialBase] = eqx.field(converter=tuple)
+
+    # def __post_init__(self) -> None:
+    #     super().__post_init__()
+    #     self._potentials: dict[str, PotentialBase]
+    #     object.__setattr__(self, "_potentials", list(self.potentials.values()))
+
+    # === Mapping ===
+
+    def __getitem__(self, key: str) -> PotentialBase:
+        return self.potentials[key]
+
+    def __iter__(self) -> iter[str]:
+        return iter(self.potentials)
+
+    def __len__(self) -> int:
+        return len(self.potentials)
+
+    # === Potential ===
 
     @jit_method()
     def energy(
@@ -29,6 +75,6 @@ class CompositePotential(PotentialBase):
         t: jt.Array,
     ) -> jt.Array:
         output = []
-        for i in range(len(self._potential_list)):
-            output.append(self._potential_list[i].energy(q, t))
+        for p in self.potentials:
+            output.append(p.energy(q, t))
         return xp.sum(xp.array(output))
